TIL/04.09/clip-matching/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers==4.26.0
pillow
python-multipart
numpy
kiwipiepy==0.20.4
python-dotenv
aiofiles
httpx